Fix qwen_image_edit_inscene: use bfloat16 dtype and cast transformer after LoRA fusion to avoid Float/BFloat16 dtype mismatch

+++ qwen_image_edit_inscene/pytorch/loader.py
@@ -90,7 +90,7 @@ class ModelLoader(ForgeModel):
         Returns:
             QwenImageTransformer2DModel with LoRA weights fused.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         pipe = QwenImageEditPipeline.from_pretrained(
             BASE_MODEL,
@@ -102,7 +102,7 @@ class ModelLoader(ForgeModel):
         )
         pipe.fuse_lora()
 
-        self._transformer = pipe.transformer
+        self._transformer = pipe.transformer.to(dtype)
         self._transformer.eval()
         return self._transformer
 
@@ -111,7 +111,7 @@ class ModelLoader(ForgeModel):
 
         Returns a dict matching QwenImageTransformer2DModel.forward() signature.
         """
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         # From model config: in_channels=64 (img_in linear input dimension)
